Change Todo list command parsing to lexical similarity.
Move functions for lexical similarity into own utilities package.
