+++ Jarvis/utilities/GeneralUtilities.py
@@ -1,3 +1,6 @@
+# -*- coding: utf-8 -*-
+
 def wordIndex(data, word):
     wordList = data.split()
     return wordList.index(word)
+
